server/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-ninja==1.1.0
django-ninja-extra==0.20.3
django-ninja-jwt==5.3.0
injector==0.21.0
install==1.3.5
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
